requirements.txt
Django==1.8
django-taggit==0.17.1
pytz==2017.3
